update version func with cmake version variable and git status detect

+++ example/Adder/example.py
@@ -6,7 +6,7 @@ if __name__ == "__main__":
 
     for i in range(6):
         dut.a.value = i * 2
-        dut.b.value = i / 4
+        dut.b.value = int(i / 4)
         dut.Step(1)
         print(dut.sum.value, dut.cout.value)
